+++ stratatools/cartridge.py
