+++ stratatools/checksum.py
