+++ stratatools/crypto.py
