+++ stratatools/formatter.py
