+++ stratatools/machine.py
