+++ stratatools/manager.py
