+++ stratatools/setupcode.py
